+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Support for controlling titles in Ext GSEA UI.
 - Added Ext gsea weighting by gene option to UI.
 - Better support for abbreviations in app info for configuring icons.
 - Sidebar layout of matcalc changed to move groups and filters around.